Move wrapping lgoic to own file
